requirements.txt
# Core Dependencies
numpy>=1.20.0
pandas>=1.3.0
opencv-python>=4.6.0
mediapipe>=0.8.10
matplotlib>=3.5.0

# Deep Learning
torch>=1.13.0
torchvision>=0.14.0
tensorflow>=2.10.0  # Optional alternative to PyTorch

# Pose Estimation
# MediaPipe is already included above

# Language Models
transformers>=4.20.0
openai>=0.27.0
sentence-transformers>=2.2.2

# ASR/TTS
librosa>=0.9.2
sounddevice>=0.4.5
# SpeechRecognition>=3.8.1  # Uncomment if needed

# Web/UI
flask>=2.0.1
streamlit>=1.10.0

# Development & Utilities
jupyter>=1.0.0
ipython>=8.0.0
pytest>=7.0.0
tqdm>=4.64.0
python-dotenv>=0.20.0

# Data Processing
scikit-learn>=1.0.2 